some review

+++ SIM-VICUS/src/core3D/Vic3DGeometryHelpers.h
@@ -156,7 +156,7 @@ void updateColors(const VICUS::Surface & s, unsigned int & currentVertexIndex,
 				  std::vector<ColorRGBA> & colorBufferData);
 
 /*! This function adds a line using addBox */
-void addLine(const IBKMK::Vector3D & startPoint, const IBKMK::Vector3D & endPoint, float width, const QColor & color,
+void addLine(const IBKMK::Vector3D & startPoint, const IBKMK::Vector3D & endPoint, double width, const QColor & color,
 			 unsigned int & currentVertexIndex, unsigned int & currentElementIndex,
 			 std::vector<Vertex> & vertexBufferData, std::vector<ColorRGBA> & colorBufferData, std::vector<GLuint> & indexBufferData);
 
@@ -166,7 +166,7 @@ void addPoint(const IBKMK::Vector3D &point, float width, const QColor &color,
 			  std::vector<Vertex> &vertexBufferData, std::vector<ColorRGBA> &colorBufferData,
 			  std::vector<GLuint> &indexBufferData);
 
-void addPolyLine(const std::vector<IBKMK::Vector3D> & polyline, bool connectEndStart, float width, const QColor & color,
+void addPolyLine(const std::vector<IBKMK::Vector3D> & polyline, bool connectEndStart, double width, const QColor & color,
 				 unsigned int & currentVertexIndex, unsigned int & currentElementIndex,
 				 std::vector<Vertex> & vertexBufferData, std::vector<ColorRGBA> & colorBufferData,
 				 std::vector<GLuint> & indexBufferData);
